refactor(apps): abstract root server (#1176)
* rebase

* pg adapter should be default with jellyfish parse

add missed newFastifyAdapter

* move start and stop to root

* overriding

* safety start

* move config to root
