+++ apps/ocean-api/testing/OceanStubServer.ts
@@ -2,14 +2,14 @@ import { MasterNodeRegTestContainer } from '@defichain/testcontainers'
 import { FastifyAdapter, NestFastifyApplication } from '@nestjs/platform-fastify'
 import { Test } from '@nestjs/testing'
 import { RootModule } from '../src/modules/RootModule'
-import { RootServer } from '../src'
+import { OceanApiServer } from '../src'
 import { ConfigModule, ConfigService } from '@nestjs/config'
 
 /**
  * Service stubs are simulations of a real service, which are used for functional testing.
  * Configures a TestingModule that is configured to connect to a provided @defichain/testcontainers.
  */
-export class OceanStubServer extends RootServer {
+export class OceanStubServer extends OceanApiServer {
   /**
    * @see PlaygroundModule
    */
