+++ apps/playground-api/testing/PlaygroundStubServer.ts
@@ -2,7 +2,7 @@ import { MasterNodeRegTestContainer } from '@defichain/testcontainers'
 import { NestFastifyApplication } from '@nestjs/platform-fastify'
 import { Test } from '@nestjs/testing'
 import { RootModule } from '../src/modules/RootModule'
-import { RootServer } from '../src'
+import { PlaygroundApiServer } from '../src'
 import { ConfigService } from '@nestjs/config'
 import { newFastifyAdapter } from '../src/Fastify'
 
@@ -10,7 +10,7 @@ import { newFastifyAdapter } from '../src/Fastify'
  * Service stubs are simulations of a real service, which are used for functional testing.
  * Configures a TestingModule that is configured to connect to a provided @defichain/testcontainers.
  */
-export class PlaygroundStubServer extends RootServer {
+export class PlaygroundStubServer extends PlaygroundApiServer {
   constructor (private readonly container: MasterNodeRegTestContainer) {
     super()
   }
